+++ debvader/detection.py
